fix(project): Switch to using yarn package manager
I have switched to using the `yarn` package manager as it is a bit
more user-friendly to use.

I also set the project to use NodeJS v20.12.2, the `.tool-versions`
file will configure the system for thos that use the `asdf` runtime
version manager. I also set the `engine` attributes in the package.json
file.