feat(devops): Add ability to build docker image
If you don't want to deploy to `vercel`, then usually you will want
to build your own docker image, so as to deploy the app to some other
container based infrastructure. I have added a Dockerfile to create
an image with the app built and running. I've also added a docker
compose file to make it easier to build/run the created image.