feat(tests): Add network request mocking
I've added MSW, which is a Mock Service Worker to intercept network
requests from the client-side. This is useful as we can:

* Mock error requests during development, to test different visualisations.
* Develop payment tools without fear of incurring costs.
* Write tests and provide mock data in the fetch response(s).

I've added the most basic fixture/handler setup that listens on a
particular wildcarded endpoint, intercepts the request and returns
a stock object as the response.

It may be that this feature is not required and is trivial to remove
since it is in one commit and can easily be reverted. It is better
to have and not need than to try install at a later date and get
inconsistent solutions.