feat(frontend): Add accessibility checking
We need to care more about accessibility checking in our apps. While
there are online tools to help, there are many and we also want to
be consistent and find issues early. Adding `Axe`, which is used by
the likes of `lighthouse`, we can see accessibility issues while we
develop. Issues found are written to the `Dev Console` and only in
development mode.