feat(project): Configure app to run on base path
Since most of the apps run on the same hostname, each app needs to
run on a different path. This is an easy way to define said path.
In the template it is set to `template-app`, for each app built from
this template, the dev just needs to configure the text for the
path.