feat(frontend): Add ability to use styled-components
The main UI component library we use is the Gov UK react components
library. This is a styled-components based library and to use it we
need access to a stylesheet manager registry, so there is some config
needing setup.

The Layout, due to the styled-compoennts, can only run on the client-side
so we need to mark the component as client-side, but then the metadata
SEO config can only be done on a server-side component, so I created
a new wrapper component to contain the client-side code and it can
take child props, so we can have server-side components wrapped in
client-side code.