feat(frontend): Add a footer
Add the shared footer, this footer should be the same for all apps.
It could be that we export the footer component to a shared library,
that way if there are any changes, it is done in the library and a
new version released, then each app just upgrades to the latest
version of the library.