feat(frontend): Add responsive Header
I've added a responsive header that while wider than a small devices
layout, shows the links just below the main header. When the app is
displayed on a small device layout, then the links are hidden and a
burger bar button is visible and when clicked displays the links in
an accordion panel.